refactor: Refactor performed by rector

+++ app/Http/Controllers/Api/ProductsMediaController.php
@@ -2,77 +2,75 @@
 
 namespace App\Http\Controllers\Api;
 
+use Illuminate\Http\JsonResponse;
 use App\Http\Controllers\Controller;
 use App\Models\Media;
 use App\Models\Product;
 
 class ProductsMediaController extends Controller
 {
-    public function destroy($id, $media_id)
+    public function destroy($id, $media_id): JsonResponse
     {
-        $media = Media::find($media_id);
+        $media = Media::query()->find($media_id);
 
-        if ($media) {
-            if ($media->model_id === $id && $media->model_type === Product::class) {
-                $media->delete();
-
-                return response()->json(['message' => 'Media deleted'], 200);
-            }
+        if ($media && ($media->model_id === $id && $media->model_type === Product::class)) {
+            $media->delete();
+            return new JsonResponse(['message' => 'Media deleted'], 200);
         }
 
-        return response()->json(['message' => 'Media not found'], 404);
+        return new JsonResponse(['message' => 'Media not found'], 404);
     }
 
-    public function destroyDraft($id)
+    public function destroyDraft($id): JsonResponse
     {
-        $media = Media::find($id);
+        $media = Media::query()->find($id);
 
         if ($media) {
             $media->delete();
 
-            return response()->json(['message' => 'Media deleted'], 200);
+            return new JsonResponse(['message' => 'Media deleted'], 200);
         }
 
-        return response()->json(['message' => 'Media not found'], 404);
+        return new JsonResponse(['message' => 'Media not found'], 404);
     }
 
-    public function draft()
+    public function draft(): JsonResponse
     {
         $file = request()->file('file');
 
         if ($file) {
             $path = $file->store('products', 'public');
-            $media = Media::create([
+            $media = Media::query()->create([
                 'model_id' => 0,
                 'model_type' => Product::class,
                 'filename' => explode('/', $path)[1],
             ]);
 
-            return response()->json(['data' => $media], 201);
+            return new JsonResponse(['data' => $media], 201);
         }
 
-        return response()->json(['message' => 'File not found'], 404);
+        return new JsonResponse(['message' => 'File not found'], 404);
     }
 
-    public function store($id)
+    public function store($id): JsonResponse
     {
-        $product = Product::find($id);
+        $product = Product::query()->find($id);
 
         if ($product) {
             $file = request()->file('file');
 
             if ($file) {
                 $path = $file->store('products', 'public');
-                $media = Media::create([
+                $media = Media::query()->create([
                     'model_id' => $product->id,
                     'model_type' => Product::class,
                     'filename' => explode('/', $path)[1],
                 ]);
 
-                return response()->json(['data' => $media], 201);
+                return new JsonResponse(['data' => $media], 201);
             }
         }
 
-        return response()->json(['message' => 'Product not found'], 404);
+        return new JsonResponse(['message' => 'Product not found'], 404);
     }
 }
